MultiModal_RAG/backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    OPENROUTER_API_KEY: str = os.getenv("OPENROUTER_API_KEY")
    LANGCHAIN_API_KEY: str = os.getenv("LANGCHAIN_API_KEY")
    CHROMA_PATH: str = os.getenv("CHROMA_PATH", "chroma_db")

    EMBEDDING_MODEL: str = "BAAI/bge-small-en-v1.5"
    LLM_MODEL: str = "mistralai/mixtral-8x7b-instruct"

settings = Settings()
